Post height fix

+++ Sources/Steepshot/Steepshot.Core/Models/Size.cs
@@ -2,7 +2,7 @@
 {
     public class Size
     {
-        public int Height { get; set; }
-        public int Width { get; set; }
+        public float Height { get; set; }
+        public float Width { get; set; }
     }
 }
